requirements.txt
dify_plugin>0.3.0
matplotlib
openai
Pillow
